refactor(Zombie): move out from PVZ.h
