+++ pvzclass/Classes/Plant.cpp
@@ -1,4 +1,5 @@
 #include "../PVZ.h"
+#include "GameObject.hpp"
 #include "Projectile.hpp"
 
 PVZ::Plant::Plant(int indexoraddress)
